feat: validation for null id

+++ controllers/delete.php
@@ -1,8 +1,9 @@
 <?php
 require_once ('../models/Pais.php');
 
-$id = $_GET['id'];
+$id = @$_GET['id'];
 
-Pais::delete(intval($id));
+if (!is_null($id))
+    Pais::delete(intval($id));
 
 header("location:/ejemplo-web");
